% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanodrop.R
\name{as_nanodrop}
\alias{as_nanodrop}
\alias{as_nanodrop.data.frame}
\title{Coerce object to nanodrop}
\usage{
as_nanodrop(x, ...)

\method{as_nanodrop}{data.frame}(x, nucleotide = NULL, is_tidy = FALSE, date = NULL, ...)
}
\arguments{
\item{x}{a \code{data.frame}}

\item{...}{additional arguments passed to their respective functions}

\item{nucleotide}{an optional character vector describing the measured nucleotide (DNA, RNA, dsDNA, dsRNA)}

\item{is_tidy}{Logical. Is \code{data} already tidy?}

\item{date}{Character coercible to a \code{Date} object via \code{lubridate::as_date}}
}
\value{
a \code{nanodrop} object
}
\description{
Coerce object to nanodrop
}
