% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr.R
\name{as_pcr}
\alias{as_pcr}
\title{Convert a scrubbed pcr object back to a pcr object}
\usage{
as_pcr(x)
}
\arguments{
\item{x}{A \code{tibble} - usually a previously scrubbed \code{pcr} object.}
}
\value{
A \code{pcr} object
}
\description{
Convert a scrubbed pcr object back to a pcr object
}
