% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incucyte.R
\name{is_incucyte}
\alias{is_incucyte}
\title{Test if an object is a incucyte object}
\usage{
is_incucyte(incucyte)
}
\arguments{
\item{incucyte}{Object to be tested}
}
\value{
logical
}
\description{
Test if an object is a incucyte object
}
