% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanodrop.R
\name{is_nanodrop}
\alias{is_nanodrop}
\title{Test if an object is a nanodrop object}
\usage{
is_nanodrop(nanodrop)
}
\arguments{
\item{nanodrop}{Object to be tested}
}
\value{
logical
}
\description{
Test if an object is a nanodrop object
}
