% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\name{layers_of_interest}
\alias{layers_of_interest}
\title{Example of variables for a set of interest}
\format{
A \code{SpatRaster} object.
}
\value{
No return value. Used with function \code{\link[terra]{rast}} to
bring raster variables to analysis.
}
\description{
A \code{SpatRaster} object representing variables in a set of interest. Variables
represent future bioclimatic variables downloaded from the WorldClim database
(\url{https://worldclim.org/}).
}
\examples{
layers_of_interest <- terra::rast(system.file("extdata",
                                              "layers_of_interest.tif",
                                              package = "mop"))

terra::plot(layers_of_interest)
}
