% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{matrix_of_interest}
\alias{matrix_of_interest}
\title{Example of matrix with variables in a set of interest}
\format{
A matrix with 723 rows and 6 columns.
}
\usage{
matrix_of_interest
}
\description{
A numeric table representing variables in a set of interest.
}
\examples{
data("matrix_of_interest", package = "mop")

head(matrix_of_interest)
}
\keyword{datasets}
