% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mop-package.R
\docType{package}
\name{mop-package}
\alias{mop}
\alias{mop-package}
\title{mop: A 'broom'-like Interface to Lab Data}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Kai Aragaki \email{adamaragaki@gmail.com} (\href{https://orcid.org/0000-0002-9458-0426}{ORCID})

}
\keyword{internal}
