% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{mop_results}
\alias{mop_results}
\alias{new_mop_results}
\title{Constructor of S3 objects of class mop_results}
\usage{
new_mop_results(summary = new("list"), mop_distances = NULL,
                mop_basic = NULL, mop_simple = NULL,
                mop_detailed = new("list"))
}
\arguments{
\item{summary}{= new("list"),}

\item{mop_distances}{= NULL,}

\item{mop_basic}{= NULL}

\item{mop_simple}{= NULL,}

\item{mop_detailed}{= new("list")}
}
\value{
An object of class \code{mop_results}.
}
\description{
Constructor of S3 objects of class mop_results
}
