% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incucyte.R
\name{new_incucyte}
\alias{new_incucyte}
\title{Constructor for a Incucyte object}
\usage{
new_incucyte(
  data = data.frame(),
  raw_data = data.frame(),
  date = lubridate::Date(),
  nucleotide = character(),
  is_tidy = logical()
)
}
\arguments{
\item{data}{\code{data.frame}, no restrictions on form.}

\item{raw_data}{\code{data.frame}, no restrictions on form. Typically represents
read-in data with no changes to it.}

\item{date}{\code{lubridate::Date} object}

\item{is_tidy}{\code{logical}. Is the \code{data.frame} provided tidy?}
}
\value{
a \code{nanodrop} object
}
\description{
Constructor for a Incucyte object
}
