% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanodrop.R
\name{new_nanodrop}
\alias{new_nanodrop}
\title{Constructor for a nanodrop object}
\usage{
new_nanodrop(
  data = data.frame(),
  raw_data = data.frame(),
  date = lubridate::Date(),
  nucleotide = character(),
  is_tidy = logical()
)
}
\arguments{
\item{data}{\code{data.frame}, no restrictions on form.}

\item{date}{\code{lubridate::Date} object}

\item{nucleotide}{\code{character}. Should likely be DNA, RNA, dsDNA, or even dsRNA.}

\item{is_tidy}{\code{logical}. Is the \code{data.frame} provided tidy?}
}
\value{
a \code{nanodrop} object
}
\description{
As this is a low-level constructor, it will let you do some un-recommended things
without warning. You have been warned.
}
