% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr.R
\name{new_pcr}
\alias{new_pcr}
\title{Constructor for a pcr object}
\usage{
new_pcr(
  data = data.frame(),
  raw_data = raw(),
  header = character(),
  footer = character(),
  date = lubridate::Date(),
  experiment_type = character(),
  wells = integer(),
  is_tidy = logical()
)
}
\arguments{
\item{data}{\code{data.frame} or \code{gp}, no restrictions on form.}

\item{raw_data}{\code{raw}, read in from the provided file with \code{readr::read_file_raw()}}

\item{header}{\code{data.frame}, the upper portion of the data before the body of
the data. Usually contains run information.}

\item{footer}{\code{data.frame}, the lower portion of the data after the body of
the data. Usually contains analysis information. May not exist.}

\item{date}{\code{lubridate::Date} object}

\item{experiment_type}{\code{character}. Could}

\item{wells}{Number of wells in plate}

\item{is_tidy}{\code{logical}. Is the \code{data.frame} provided tidy?}
}
\value{
A \code{pcr} object
}
\description{
Constructor for a pcr object
}
