% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synergy2.R
\name{new_synergy2}
\alias{new_synergy2}
\title{Constructor for a synergy2 object}
\usage{
new_synergy2(data = data.frame(), wavelengths = numeric(), metadata = list())
}
\arguments{
\item{data}{a \code{data.frame}, no restrictions on form.}

\item{wavelengths}{a num}

\item{metadata}{a \code{list}, no restrictions on form}
}
\value{
a \code{synergy2} object
}
\description{
Constructor for a synergy2 object
}
