% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/out_range.R
\name{out_range}
\alias{out_range}
\title{Values outside ranges of reference conditions}
\usage{
out_range(m_matrix, g_matrix, type = "basic")
}
\arguments{
\item{m_matrix}{matrix of variables representing the set of conditions to be
used as reference. Each column represents a variable.}

\item{g_matrix}{matrix of variables representing the set of conditions to be
compared against the reference conditions (where conditions outside range
are to be detected). Each column represents a variable. Variable names must
match those in \code{m_matrix}.}

\item{type}{(character) type of identification to be performed. Options
are: "basic", "simple", and "detailed". Default = "basic".}
}
\description{
Options to identify which values in a set of conditions of interest
(\code{g_matrix}) are outside the range of a set of conditions of
reference (\code{m_matrix}).
}
