% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pad_zero}
\alias{pad_zero}
\title{Add leading 0 to sample names}
\usage{
pad_zero(sample_names)
}
\arguments{
\item{sample_names}{a vector of sample names}
}
\value{
A sample name with up to one zero padded
}
\description{
Add leading 0 to sample names
}
\examples{
c("blueberry", "Sample 1", "Sample 10", "sample 2", "john", "larry", "toxic waste", "Sample 02") |>
  pad_zero()

# pad_zero will not pad if it doesn't need to:
c("Sample 1", "Sample 2", "Sample 9") |> pad_zero()

# pad_zero is case sensitive, since these will be coming off a machine in a standardized format

c("sample 10", "sample 1") |> pad_zero()
c("Sample 10", "Sample 1") |> pad_zero()
}
