% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_generics.R, R/spectramax.R, R/synergy2.R
\name{plate_data}
\alias{plate_data}
\alias{plate_data<-}
\alias{plate_data.spectramax}
\alias{plate_data<-.spectramax}
\alias{plate_data.synergy2}
\alias{plate_data<-.synergy2}
\title{Get or set an object's plate data}
\usage{
plate_data(x, ...)

plate_data(x) <- value

\method{plate_data}{spectramax}(x, ...)

\method{plate_data}{spectramax}(x) <- value

\method{plate_data}{synergy2}(x, ...)

\method{plate_data}{synergy2}(x) <- value
}
\arguments{
\item{x}{Object to extract plate data from}

\item{...}{Arguments to be passed to their respective methods}

\item{value}{Value to set the plate to.}
}
\value{
A \code{gp} object
}
\description{
Get or set an object's plate data
}
