% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print}
\alias{print}
\alias{print.mop_results}
\alias{print,mop_results}
\title{Print a short version of elements in mop objects}
\usage{
\method{print}{mop_results}(x, ...)
}
\arguments{
\item{x}{object of class mop_results.}

\item{...}{further arguments to be passed to or from other methods. Ignored
in this function.}
}
\value{
A short description of objects in the console.
}
\description{
Print a short version of elements in mop objects
}
