% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incucyte_read.R
\name{read_incucyte}
\alias{read_incucyte}
\title{Read an Incucyte File}
\usage{
read_incucyte(data, platemap)
}
\arguments{
\item{data}{Plotting data output from the Incucyte analysis}

\item{platemap}{.PlateMap file output from the platemap maker}
}
\description{
Read an Incucyte File
}
