% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_read.R
\name{read_pcr}
\alias{read_pcr}
\title{Read in a QuantStudio pcr file}
\usage{
read_pcr(path)
}
\arguments{
\item{path}{path to a QuantStudio .xls}
}
\value{
a \code{pcr} object
}
\description{
Read in a QuantStudio pcr file
}
