% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synergy2_read.R
\name{read_synergy2}
\alias{read_synergy2}
\title{Read and tidy a Synergy2 file}
\usage{
read_synergy2(path)
}
\arguments{
\item{path}{Path to the Synergy2 .txt file}
}
\value{
A list, with the following form: \describe{
\item{data}{a \code{tibble} with all well-level experimental data combined}
\item{wavelengths}{a numeric vector containing the wavelengths read}
\item{metadata}{A \code{list} containing residual, non-well data}
}
}
\description{
Read and tidy a Synergy2 file
}
\examples{
system.file("extdata", "bca-assay-all-export.txt", package = "mop") |>
  read_synergy2()
}
