% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\name{reference_layers}
\alias{reference_layers}
\title{Example of variables for a set of reference}
\format{
A SpatRaster object.
}
\value{
No return value, used with function \code{\link[terra]{rast}} to
bring raster variables used in analysis.
}
\description{
A SpatRaster object representing variables in a set of reference.
}
\examples{
reference_layers <- terra::rast(system.file("extdata", "reference_layers.tif",
                                            package = "mop"))

terra::plot(reference_layers)
}
