% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{reference_matrix}
\alias{reference_matrix}
\title{Example of matrix with variables in a set of reference}
\format{
A matrix with 723 rows and 6 columns.
}
\usage{
reference_matrix
}
\description{
A numeric table representing variables in a set of reference.
}
\examples{
data("reference_matrix", package = "mop")

head(reference_matrix)
}
\keyword{datasets}
