% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nanodrop_scrub.R
\name{scrub.nanodrop}
\alias{scrub.nanodrop}
\title{Convert a nanodrop object into a tibble}
\usage{
\method{scrub}{nanodrop}(x, raw = FALSE)
}
\arguments{
\item{x}{a \code{nanodrop} object}

\item{raw}{Should the data be pulled from \code{x$raw_data}?}
}
\value{
a \code{tibble}
}
\description{
Convert a nanodrop object into a tibble
}
