% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_scrub.R
\name{scrub.pcr}
\alias{scrub.pcr}
\title{Convert a pcr object into a tibble}
\usage{
\method{scrub}{pcr}(x, include_header = FALSE, ...)
}
\arguments{
\item{x}{a \code{pcr} object}

\item{include_header}{Should the data from \code{x$header} be included?}

\item{...}{Unused}
}
\value{
a \code{tibble}
}
\description{
Convert a pcr object into a tibble
}
