% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrub.R
\name{scrub}
\alias{scrub}
\title{Convert a lab object to a tidy tibble}
\usage{
scrub(x, ...)
}
\arguments{
\item{x}{object to be converted into a \code{tibble}}

\item{...}{additional arguments to pass on to the respective object's class method}
}
\value{
a \code{tibble}
}
\description{
Objects are a convenient way to represent lab data as it allows downstream
functions to interact with them in specific ways depending on their type.
Sometimes, however, it is more convenient to work with a flat \code{tibble}.
\code{scrub} converts a lab object into a \code{tibble}
}
