% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectramax_scrub.R
\name{scrub.spectramax}
\alias{scrub.spectramax}
\title{Convert a spectramax object into a tibble}
\usage{
\method{scrub}{spectramax}(x, n = 1, ...)
}
\arguments{
\item{x}{a \code{spectramax} object}

\item{n}{List item number to select data from}

\item{...}{Unused}
}
\value{
a \code{tibble}
}
\description{
Convert a spectramax object into a tibble
}
