% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/incucyte_tidy.R
\name{tidy_lab.incucyte_data}
\alias{tidy_lab.incucyte_data}
\title{Tidy an incucyte object}
\usage{
\method{tidy_lab}{incucyte_data}(x, force_tidy = FALSE, ...)
}
\arguments{
\item{x}{An \code{incucyte} object to tidy}

\item{force_tidy}{\code{logical}. Should the tidying take place, even if the
\code{is_tidy} attribute is \code{TRUE}?}

\item{...}{Unused}
}
\value{
an \code{incucyte} object
}
\description{
Tidy an incucyte object
}
\details{
This function:
}
