% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectramax_tidy.R, R/tidy_lab.R
\name{tidy_lab.spectramax}
\alias{tidy_lab.spectramax}
\alias{tidy_lab}
\title{Tidy a lab object}
\usage{
\method{tidy_lab}{spectramax}(x, ...)

tidy_lab(x, ...)
}
\arguments{
\item{x}{An object to tidy}

\item{...}{Arguments passed to their respective methods}
}
\description{
Tidy a lab object
}
