% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mop_distance.R
\name{mop_distance}
\alias{mop_distance}
\title{MOP distance calculation}
\usage{
mop_distance(m_matrix, g_matrix, distance = "euclidean", percentage = 1,
             comp_each = 2000, parallel = FALSE, n_cores = NULL,
             progress_bar = TRUE)
}
\arguments{
\item{m_matrix}{matrix of variables representing the set of conditions to be
used as reference. Each column represents a variable.}

\item{g_matrix}{matrix of variables representing the set of conditions to be
compared against the reference conditions (where distances are to be
calculated). Each column represents a variable. Variable names must match
those in \code{m_matrix}.}

\item{distance}{(character) one of the options: "euclidean" or "mahalanobis".
Default = "euclidean".}

\item{percentage}{(numeric) percentage of points of m (the closest ones)
used to derive mean environmental distances to each g point. Default = 1.}

\item{comp_each}{(numeric) number of points of the g matrix to be used for
distance calculations at a time (default = 2000). Increasing this number
requires more RAM.}

\item{parallel}{(logical) if TRUE, calculations will be performed in parallel
using \code{n_cores} of the computer. Using this option will speed up the
analysis  but will demand more RAM. Default = FALSE.}

\item{n_cores}{(numeric) number of cores to be used in parallel processing.
Default = NULL, in which case all CPU cores on current host - 1 will be used.}

\item{progress_bar}{(logical) whether to show a progress bar for calculations,
default = TRUE. Valid when calculation are not run in parallel.}
}
\description{
Calculates distances from each of the points of interest in \code{g_matrix}
to a defined percentage of the reference conditions in \code{m_matrix}.
}
