% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{mop_results}
\alias{mop_results}
\alias{new_mop_results}
\title{Constructor of S3 objects of class mop_results}
\usage{
new_mop_results(summary = new("list"), mop_distances = NULL,
                mop_basic = NULL, mop_simple = NULL,
                mop_detailed = new("list"))
}
\arguments{
\item{summary}{a list with a summary of the data and parameters used in
analysis. Default = empty list.}

\item{mop_distances}{a \code{SpatRaster} or numeric vector of distances from the
set of conditions of reference to the set of conditions of interest.
Default = NULL.}

\item{mop_basic}{a \code{SpatRaster} or numeric vector showing conditions in the set
of interest outside the ranges in the reference set. The value \code{1}
indicates conditions outside one or more ranges. Default = NULL.}

\item{mop_simple}{a \code{SpatRaster} or numeric vector showing conditions in the
set of interest outside the ranges in the reference set. Values indicate how
many variables are outside reference ranges. Default = NULL.}

\item{mop_detailed}{a list with a detailed representation of mop results
in conditions outside the range of reference. Default = empty list.}
}
\value{
An object of class \code{mop_results}.
}
\description{
Constructor of S3 objects of class mop_results
}
