% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr.R
\name{new_pcr}
\alias{new_pcr}
\title{Constructor for a pcr object}
\usage{
new_pcr(
  data = data.frame(),
  raw_data = data.frame(),
  header = data.frame(),
  footer = data.frame(),
  date = lubridate::Date(),
  experiment_type = character(),
  is_tidy = logical()
)
}
\arguments{
\item{data}{\code{data.frame}, no restrictions on form.}

\item{raw_data}{\code{data.frame}, no restrictions on form. Typically represents
read-in data with no changes to it.}

\item{header}{\code{data.frame}, the upper portion of the data before the body of
the data. Usually contains run information.}

\item{footer}{\code{data.frame}, the lower portion of the data after the body of
the data. Usually contains analysis information. May not exist.}

\item{date}{\code{lubridate::Date} object}

\item{experiment_type}{\code{character}. Could}

\item{is_tidy}{\code{logical}. Is the \code{data.frame} provided tidy?}
}
\value{

}
\description{
Constructor for a pcr object
}
