% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectramax.R
\name{new_spectramax}
\alias{new_spectramax}
\title{Constructor for a spectramax object}
\usage{
new_spectramax(
  data = data.frame(),
  raw_data = raw(),
  date = lubridate::Date(),
  experiment_type = character(),
  is_tidy = logical()
)
}
\arguments{
\item{data}{a \code{data.frame}, no restrictions on form.}

\item{raw_data}{\code{raw}, read in from the provided file with \code{readr::read_file_raw()}}

\item{date}{\code{lubridate::Date} object}

\item{experiment_type}{\code{character}, no restrictions on form}

\item{is_tidy}{\code{logical}. Are the \code{data} provided tidy?}
}
\value{
a \code{spectramax} object
}
\description{
Constructor for a spectramax object
}
