% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary}
\alias{summary}
\alias{summary.mop_results}
\alias{summary,mop_results}
\title{Summary of attributes and results}
\usage{
\method{summary}{mop_results}(object, ...)
}
\arguments{
\item{object}{object of class mop_results.}

\item{...}{additional arguments affecting the summary produced. Ignored in
this function.}
}
\value{
A printed summary.
}
\description{
Summary of attributes and results
}
