% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcr_tidy.R
\name{tidy_lab.pcr}
\alias{tidy_lab.pcr}
\title{Tidy a PCR object}
\usage{
\method{tidy_lab}{pcr}(x, usr_standards = NULL, pad_zero = FALSE, ...)
}
\arguments{
\item{x}{a \code{pcr} object}

\item{usr_standards}{Custom supplied standards}

\item{pad_zero}{Should, say, Sample 1 become Sample 01?}

\item{...}{}
}
\description{
Tidy a PCR object
}
